Add RecipeDetailActivity

+++ app/src/main/java/br/com/gabriellferreira/recipelist/domain/model/Recipe.kt
@@ -3,7 +3,7 @@ package br.com.gabriellferreira.recipelist.domain.model
 class Recipe(
     val id: String,
     val title: String,
-    val calories: Double,
+    val calories: Int,
     val description: String,
     val chef: String,
     val tags: List<String>,
